backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
httpx
sqlmodel
python-dotenv
cryptography
routeros-api
